requirements.txt
aiohttp>=3.9.1
pytest>=7.4.3
pytest-asyncio>=0.23.2
python-dotenv>=1.0.0
tenacity>=8.2.3
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
ebooklib>=0.18.0
beautifulsoup4>=4.12.0
pandoc>=2.3
brotli>=1.1.0
